Merge Upstream Master

+++ scilab/modules/scicos/sci_gateway/cpp/sci_scicos_debug.cpp
@@ -1,24 +1,24 @@
 /*  Scicos
-*
-*  Copyright (C) 2015 - Scilab Enterprises - Paul Bignier
-*  Copyright (C) INRIA - Allan CORNET
-*
-* This program is free software; you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation; either version 2 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License
-* along with this program; if not, write to the Free Software
-* Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-*
-* See the file ./license.txt
-*/
+ *
+ *  Copyright (C) 2015 - Scilab Enterprises - Paul Bignier
+ *  Copyright (C) INRIA - Allan CORNET
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ *
+ * See the file ./license.txt
+ */
 
 #include <string>
 #include <cmath>
@@ -82,7 +82,7 @@ types::Function::ReturnValue sci_scicos_debug(types::typed_list &in, int _iRetCo
             return types::Function::Error;
         }
 
-        C2F(cosdebug).cosd = pIn->getFirst();
+        C2F(cosdebug).cosd = (int)(pIn->getFirst());
     }
 
     return types::Function::OK;
